fixed some conversion problems when compiling without vectorization support
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2927 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/datadriven/basis/modlinear/operation/OperationMultipleEvalIterativeSPHybridX86SimdOCLModLinear.cpp
@@ -457,7 +457,7 @@ namespace sg {
                 } else {
                   float eval = ((ptrLevel[(j * dims) + d]) * (ptrData[(i * dims) + d]));
                   float index_calc = eval - (ptrIndex[(j * dims) + d]);
-                  float abs = fabs(index_calc);
+                  float abs = static_cast<float>(fabs(index_calc));
                   float last = 1.0f - abs;
                   float localSupport = std::max<float>(last, 0.0f);
                   curSupport *= localSupport;
@@ -899,7 +899,7 @@ namespace sg {
                 } else {
                   float eval = ((ptrLevel[(j * dims) + d]) * (ptrData[(i * dims) + d]));
                   float index_calc = eval - (ptrIndex[(j * dims) + d]);
-                  float abs = fabs(index_calc);
+                  float abs = static_cast<float>(fabs(index_calc));
                   float last = 1.0f - abs;
                   float localSupport = std::max<float>(last, 0.0f);
                   curSupport *= localSupport;
